parsing 테스트 전

+++ libft/list_getter.c
@@ -6,13 +6,13 @@
 /*   By: sejjeong <sejjeong@student.42gyeongsan>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 16:22:44 by sejjeong          #+#    #+#             */
-/*   Updated: 2024/11/05 16:24:02 by sejjeong         ###   ########.fr       */
+/*   Updated: 2024/11/08 16:58:05 by sejjeong         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "array_list.h"
 
-void	*get_element_or_null_in_list(t_array_list *list, int index)
+void	*get_element_or_null_in_list(t_array_list *list, size_t index)
 {
 	if (index >= list->get_count(list))
 	{
